requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
python-multipart==0.0.20
openai==2.31.0
pypdf==6.0.0
numpy==2.3.1
sentence-transformers==5.1.0
